abc132

+++ src/abc132/blue_and_red_balls.h
@@ -1,2 +1,2 @@
 
-std::vector<int> solve_blue_and_red_balls(int n, int k);
+std::vector<long long> solve_blue_and_red_balls(int n, int k);
